refactor(spec): hoist target-verify behind SpecTarget; arch-free NgramSpeculator
Make the model-free n-gram speculator 100% arch-agnostic by moving the target
verify + recurrent-state fixup behind the SpecTarget trait. No behavior change:
n-gram 9B greedy output is byte-identical before/after (md5 1d414b48).

- spec.rs: add SpecTarget::{new_spec_scratch, spec_advance, verify_block,
  commit_prefix, eos_token, ctx_capacity} + the erased SpecScratch trait (with an
  explicit free(self: Box<Self>, gpu) — GpuTensor/DeviceBuffer have no Drop here)
  + SpecAdvance. Document the save-first (verify_block) / restore-on-partial
  (commit_prefix) recurrent-state contract.
- Move NgramCache/PldMatcher/PldMatch (pure CPU) to hipfire_runtime::spec;
  re-export from hipfire-arch-qwen35::speculative so spec_step_dflash + the demo
  keep compiling unchanged.
- Relocate NgramSpeculator loader -> hipfire_runtime::spec_ngram, stripped of all
  qwen35 imports: step() is pure policy (build_draft + accept) over the trait;
  verify scratch is built lazily on first prefill via new_spec_scratch and freed
  in free().
- qwen35 spec_impl.rs: Qwen35SpecScratch (VerifyScratch + DeltaNetSnapshot + s_ef
  backup + num_extract=0 ring) + the SpecTarget verify impl — behavior-preserving
  port of the old inline logic (full-accept skip / partial restore+s_ef+batched
  replay).
- build_speculator becomes arch-typeless (arch_id, dflash, eviction_is_none,
  ctx_capacity); lib.rs call site simplified.

Validation: n-gram 9B byte-identical (md5 match); serve-multiturn-gate PASS
(AR + DFlash arms); build/clippy/per-file-rustfmt clean. DFlash/DDTree/MTP
untouched (still call the concrete fns). Change 1 of 2; llama impl + daemon
wiring follow.
