+++ crates/hipfire-arch-qwen35/src/speculative.rs
@@ -1945,183 +1945,11 @@ fn sample_residual(p_target: &[f32], p_draft: &[f32], u: f32) -> u32 {
     (p_target.len() - 1) as u32
 }
 
-/// Rolling bigram n-gram cache. Keyed by the last two committed tokens
-/// `(a, b)`; value is a small map from possible next-token to count.
-///
-/// Populated incrementally from the committed output stream. Used as a
-/// "free" second opinion on top of the DFlash draft: if the cache has
-/// seen a (a, b) → c transition with high enough count, and the DFlash
-/// draft proposed something else at that position, the n-gram's `c`
-/// often turns out to match the target's argmax.
-///
-/// Scales: the cache size is bounded by the number of distinct bigrams
-/// in the committed output — typically a few hundred per session, so
-/// no eviction policy needed.
-pub struct NgramCache {
-    /// `(a, b) → { next: count, ... }` with the next-token histogram.
-    pub bigram: std::collections::HashMap<(u32, u32), std::collections::HashMap<u32, u32>>,
-    /// Minimum count before we trust the prediction. Smaller = more
-    /// aggressive (more overrides), larger = more conservative. 3 is a
-    /// reasonable default on hot-loop code / repetitive text.
-    pub min_count: u32,
-}
-
-impl NgramCache {
-    pub fn new(min_count: u32) -> Self {
-        Self {
-            bigram: std::collections::HashMap::new(),
-            min_count,
-        }
-    }
-
-    /// Record the triple `(a, b) → c` in the cache.
-    #[inline]
-    pub fn observe(&mut self, a: u32, b: u32, c: u32) {
-        *self.bigram.entry((a, b)).or_default().entry(c).or_insert(0) += 1;
-    }
-
-    /// Predict `c` from last-two `(a, b)` if the max-count next-token
-    /// reaches `min_count`. Returns (token, count).
-    #[inline]
-    pub fn predict(&self, a: u32, b: u32) -> Option<(u32, u32)> {
-        let map = self.bigram.get(&(a, b))?;
-        let (&tok, &cnt) = map.iter().max_by_key(|(_, &c)| c)?;
-        if cnt >= self.min_count {
-            Some((tok, cnt))
-        } else {
-            None
-        }
-    }
-
-    /// Record every consecutive triple in a slice of committed tokens.
-    /// Caller supplies the full token stream; this walks it in-place.
-    pub fn observe_many(&mut self, tokens: &[u32]) {
-        if tokens.len() >= 3 {
-            for w in tokens.windows(3) {
-                self.observe(w[0], w[1], w[2]);
-            }
-        }
-    }
-}
-
-/// Prompt Lookup Decoding (Saxena 2023): training-free deterministic draft
-/// built from context suffix self-match. If the last N tokens of context
-/// appeared earlier in context, the tokens that followed that earlier
-/// occurrence are a high-quality continuation guess.
-///
-/// Used as the draft source in Goose bypass mode (Jin et al. 2026,
-/// arXiv:2604.02047 §4.3): PLD-matched tokens have 2–18× higher acceptance
-/// than bigram (TR) tokens (median 6× across 5 models × 5 benchmarks).
-/// When PLD confidence is high, the spine — a deep linear chain of
-/// PLD-matched tokens — is verified in one target forward pass without
-/// tree construction. That's exactly what we need on Qwen3.5 hybrid
-/// (24 DeltaNet + 8 FullAttention): linear verify sidesteps the
-/// state-forking problem that tree verify imposes on recurrent LA layers.
-pub struct PldMatcher {
-    /// n-gram suffix lengths to try, longest first. Paper uses {5,4,3}.
-    /// Longer matches are more selective; if the longest fails we fall
-    /// back to shorter. Order matters: we return the first (longest) hit.
-    pub ngram_lens: Vec<usize>,
-    /// Hard cap on spine length. Paper uses 8 — sufficient for typical
-    /// block sizes and avoids running off the end of a match into drift.
-    pub max_extract: usize,
-    /// Minimum extracted length to count as a usable spine. Very short
-    /// spines aren't worth the PLD path (bigram covers 1-token lookahead
-    /// at lower risk); require at least this many continuation tokens.
-    pub min_extract: usize,
-}
-
-impl Default for PldMatcher {
-    fn default() -> Self {
-        Self {
-            ngram_lens: vec![5, 4, 3],
-            max_extract: 8,
-            min_extract: 3,
-        }
-    }
-}
-
-/// Result of a successful PLD lookup.
-#[derive(Debug, Clone)]
-pub struct PldMatch {
-    /// The extracted spine (continuation tokens after the matched suffix).
-    pub tokens: Vec<u32>,
-    /// The suffix length that produced this match (the longest that hit).
-    pub n: usize,
-    /// Number of tried n-gram lengths that agreed on `tokens[0]`. Paper
-    /// §4.3 uses this as part of the bypass-mode confidence signal;
-    /// higher consensus = more reliable spine. Ranges 1..=ngram_lens.len().
-    pub consensus: usize,
-}
-
-impl PldMatcher {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Find a spine continuation for `context`. Returns `None` if no tried
-    /// n-gram length produces a match of length ≥ `self.min_extract`.
-    ///
-    /// For each n in `self.ngram_lens`: take the last-n tokens as the
-    /// suffix, search for its last occurrence earlier in context, and
-    /// extract the `max_extract` tokens that followed it (stopping before
-    /// the suffix itself so we don't include tokens that would be about
-    /// to be re-predicted). Returns the longest-n match with a usable
-    /// spine; consensus counts how many alternate n's produced the same
-    /// first continuation token.
-    pub fn lookup(&self, context: &[u32]) -> Option<PldMatch> {
-        if self.ngram_lens.is_empty() {
-            return None;
-        }
-        // Per-n continuation, collected to compute consensus across lengths.
-        let mut firsts: Vec<u32> = Vec::with_capacity(self.ngram_lens.len());
-        let mut best: Option<(usize, Vec<u32>)> = None; // (n, spine)
-        for &n in &self.ngram_lens {
-            if context.len() <= n {
-                continue;
-            }
-            let suffix_start = context.len() - n;
-            let suffix = &context[suffix_start..];
-            let haystack = &context[..suffix_start];
-            if haystack.len() < n {
-                continue;
-            }
-            // Last occurrence (freshest) of `suffix` in `haystack`.
-            let mut found: Option<usize> = None;
-            for i in (0..=haystack.len() - n).rev() {
-                if &haystack[i..i + n] == suffix {
-                    found = Some(i);
-                    break;
-                }
-            }
-            let start = match found {
-                Some(s) => s,
-                None => continue,
-            };
-            let cont_start = start + n;
-            let cont_end = (cont_start + self.max_extract).min(suffix_start);
-            if cont_end <= cont_start {
-                continue;
-            }
-            let spine: Vec<u32> = context[cont_start..cont_end].to_vec();
-            if spine.len() < self.min_extract {
-                continue;
-            }
-            firsts.push(spine[0]);
-            if best.is_none() {
-                best = Some((n, spine));
-            }
-        }
-
-        let (n, tokens) = best?;
-        let consensus = firsts.iter().filter(|&&t| t == tokens[0]).count();
-        Some(PldMatch {
-            tokens,
-            n,
-            consensus,
-        })
-    }
-}
+// `NgramCache` / `PldMatcher` / `PldMatch` moved to `hipfire_runtime::spec` so
+// the arch-generic `NgramSpeculator` can use them without an arch-crate
+// dependency. Re-exported here so `spec_step_dflash` (and the dflash_spec_demo)
+// keep referring to them unqualified.
+pub use hipfire_runtime::spec::{NgramCache, PldMatch, PldMatcher};
 
 /// Small, fast RNG for per-cycle sampling u ∈ [0, 1). Xorshift64*; deterministic
 /// given the seed, cheap enough to inline into the B-rejection loop.
