+++ crates/hipfire-runtime/src/lib.rs
@@ -42,6 +42,7 @@ pub mod paro;
 pub mod safetensors_source;
 pub mod sampler;
 pub mod spec;
+pub mod spec_ngram;
 pub mod tp_shard;
 #[cfg(feature = "deltanet")]
 pub mod triattn;
