+++ crates/hipfire-runtime/src/spec.rs
@@ -90,6 +90,109 @@ pub trait SpecTarget {
     /// offset — used by the daemon's mid-generation abort path in place of its
     /// current inline memset loop.
     fn reset_recurrent(&mut self, gpu: &mut Gpu);
+
+    // ── Arch-generic speculation primitives ─────────────────────────────────
+    //
+    // These let a *model-free* speculator (n-gram / PLD) drive any arch's target
+    // without knowing its internals: the target owns ALL verify mechanics (the
+    // batched forward, the per-position lm_head, the recurrent snapshot/rewind,
+    // and the arch-specific scratch), while the speculator owns only policy
+    // (drafting + acceptance). The arch-specific verify scratch is created by the
+    // target via [`new_spec_scratch`](Self::new_spec_scratch) and handed back on
+    // every call as an erased `&mut dyn SpecScratch`, so no arch type leaks into
+    // the speculator and the speculator owns the scratch's lifetime.
+
+    /// Allocate arch-specific verify scratch sized to `block_size` (the max
+    /// speculation window). The speculator owns the returned box for its lifetime
+    /// and frees it via [`SpecScratch::free`].
+    fn new_spec_scratch(
+        &mut self,
+        gpu: &mut Gpu,
+        block_size: usize,
+    ) -> Result<Box<dyn SpecScratch>, String>;
+
+    /// Advance the target over `tokens` from absolute `start_pos` (chunked,
+    /// abortable), returning the greedy argmax at the LAST position. `reset`
+    /// zeroes recurrent + KV state first (cache-miss prefill); `false` continues
+    /// from the current state (cache-hit suffix, or the partial-accept replay).
+    fn spec_advance(
+        &mut self,
+        gpu: &mut Gpu,
+        tokens: &[u32],
+        start_pos: usize,
+        reset: bool,
+        abort: &dyn Fn() -> bool,
+    ) -> Result<SpecAdvance, String>;
+
+    /// Run the target over `block` at absolute `position`, returning the greedy
+    /// argmax at each of the `block.len()` positions (`argmax[i]` is the target's
+    /// next-token prediction after consuming `block[0..=i]`). Leaves target state
+    /// advanced by `block.len()`.
+    ///
+    /// CONTRACT: this MUST first snapshot whatever recurrent state
+    /// [`commit_prefix`](Self::commit_prefix) needs to rewind (e.g. the DeltaNet
+    /// S/conv state AND the Q8 error-feedback residual) INTO `scratch`, *before*
+    /// running the forward that advances it. Stateless (pure-attention) arches
+    /// snapshot nothing.
+    fn verify_block(
+        &mut self,
+        gpu: &mut Gpu,
+        block: &[u32],
+        position: usize,
+        scratch: &mut dyn SpecScratch,
+    ) -> Result<Vec<u32>, String>;
+
+    /// Fix target state to reflect exactly the committed prefix
+    /// `block[..accept_len + 1]` (after [`verify_block`](Self::verify_block)
+    /// over-advanced it by `block.len()`). Cases:
+    /// - full accept (`accept_len == block.len() - 1`): no-op — verify already
+    ///   left state at the right position;
+    /// - recurrent + partial: restore the snapshot saved in `scratch` (incl. the
+    ///   s_ef residual) and replay `block[..accept_len + 1]` with the SAME batched
+    ///   forward `verify_block` used (numerics must match the accepted argmax);
+    /// - stateless + partial: no-op — the accepted-prefix KV the verify wrote is
+    ///   already correct, and the rejected tail is overwritten by the next verify.
+    fn commit_prefix(
+        &mut self,
+        gpu: &mut Gpu,
+        block: &[u32],
+        accept_len: usize,
+        position: usize,
+        scratch: &mut dyn SpecScratch,
+    ) -> Result<(), String>;
+
+    /// The target's EOS token id (for the daemon's decode-loop terminator check).
+    fn eos_token(&self) -> u32;
+
+    /// The target's usable context capacity (decode-loop overflow guard).
+    fn ctx_capacity(&self) -> usize;
+}
+
+/// Erased, arch-specific verify scratch owned by a model-free speculator.
+///
+/// The concrete scratch (qwen35: `VerifyScratch` + `DeltaNetSnapshot` + s_ef
+/// backup + hidden ring; llama: lm_head/argmax buffers) is crate-local to each
+/// arch and this crate depends on none of them — so it is threaded through the
+/// trait erased, mirroring [`SpecGrammar`]. The arch's [`SpecTarget`] impl
+/// recovers it via `scratch.as_any_mut().downcast_mut::<T>()`.
+pub trait SpecScratch {
+    /// Downcast hook for the owning [`SpecTarget`] impl.
+    fn as_any_mut(&mut self) -> &mut dyn std::any::Any;
+
+    /// Release all GPU buffers the scratch owns. Explicit because `GpuTensor` /
+    /// `DeviceBuffer` have no `Drop` in this codebase — a bare `drop` of the box
+    /// would orphan device memory. Called from the speculator's `free`.
+    fn free(self: Box<Self>, gpu: &mut Gpu);
+}
+
+/// Outcome of [`SpecTarget::spec_advance`].
+#[derive(Debug, Clone)]
+pub enum SpecAdvance {
+    /// Advanced to the end; `last_argmax` is the greedy token at the final
+    /// position (the first decode seed on prefill; ignored on replay).
+    Ready { last_argmax: u32 },
+    /// Client cancelled mid-advance; the target reset its own state.
+    Aborted,
 }
 
 /// Erased grammar-mask interface for tool-call-constrained spec-decode.
@@ -228,6 +331,190 @@ pub trait Speculator {
     /// request against a greedy-only drafter, routing it to the AR path.
     fn requires_greedy(&self) -> bool {
         true
+    }
+}
+
+// ─── Model-free drafting sources (arch-agnostic, pure CPU) ──────────────────
+//
+// Moved here from `hipfire-arch-qwen35::speculative` so the arch-generic
+// `NgramSpeculator` can use them without an arch-crate dependency. qwen35's
+// `spec_step_dflash` still uses them via a `pub use` re-export in that crate.
+
+/// Rolling bigram n-gram cache. Keyed by the last two committed tokens
+/// `(a, b)`; value is a small map from possible next-token to count.
+///
+/// Populated incrementally from the committed output stream. Used as a
+/// "free" second opinion on top of the DFlash draft: if the cache has
+/// seen a (a, b) → c transition with high enough count, and the DFlash
+/// draft proposed something else at that position, the n-gram's `c`
+/// often turns out to match the target's argmax.
+///
+/// Scales: the cache size is bounded by the number of distinct bigrams
+/// in the committed output — typically a few hundred per session, so
+/// no eviction policy needed.
+pub struct NgramCache {
+    /// `(a, b) → { next: count, ... }` with the next-token histogram.
+    pub bigram: std::collections::HashMap<(u32, u32), std::collections::HashMap<u32, u32>>,
+    /// Minimum count before we trust the prediction. Smaller = more
+    /// aggressive (more overrides), larger = more conservative. 3 is a
+    /// reasonable default on hot-loop code / repetitive text.
+    pub min_count: u32,
+}
+
+impl NgramCache {
+    pub fn new(min_count: u32) -> Self {
+        Self {
+            bigram: std::collections::HashMap::new(),
+            min_count,
+        }
+    }
+
+    /// Record the triple `(a, b) → c` in the cache.
+    #[inline]
+    pub fn observe(&mut self, a: u32, b: u32, c: u32) {
+        *self.bigram.entry((a, b)).or_default().entry(c).or_insert(0) += 1;
+    }
+
+    /// Predict `c` from last-two `(a, b)` if the max-count next-token
+    /// reaches `min_count`. Returns (token, count).
+    #[inline]
+    pub fn predict(&self, a: u32, b: u32) -> Option<(u32, u32)> {
+        let map = self.bigram.get(&(a, b))?;
+        let (&tok, &cnt) = map.iter().max_by_key(|(_, &c)| c)?;
+        if cnt >= self.min_count {
+            Some((tok, cnt))
+        } else {
+            None
+        }
+    }
+
+    /// Record every consecutive triple in a slice of committed tokens.
+    /// Caller supplies the full token stream; this walks it in-place.
+    pub fn observe_many(&mut self, tokens: &[u32]) {
+        if tokens.len() >= 3 {
+            for w in tokens.windows(3) {
+                self.observe(w[0], w[1], w[2]);
+            }
+        }
+    }
+}
+
+/// Prompt Lookup Decoding (Saxena 2023): training-free deterministic draft
+/// built from context suffix self-match. If the last N tokens of context
+/// appeared earlier in context, the tokens that followed that earlier
+/// occurrence are a high-quality continuation guess.
+///
+/// Used as the draft source in Goose bypass mode (Jin et al. 2026,
+/// arXiv:2604.02047 §4.3): PLD-matched tokens have 2–18× higher acceptance
+/// than bigram (TR) tokens (median 6× across 5 models × 5 benchmarks).
+/// When PLD confidence is high, the spine — a deep linear chain of
+/// PLD-matched tokens — is verified in one target forward pass without
+/// tree construction. That's exactly what we need on Qwen3.5 hybrid
+/// (24 DeltaNet + 8 FullAttention): linear verify sidesteps the
+/// state-forking problem that tree verify imposes on recurrent LA layers.
+pub struct PldMatcher {
+    /// n-gram suffix lengths to try, longest first. Paper uses {5,4,3}.
+    /// Longer matches are more selective; if the longest fails we fall
+    /// back to shorter. Order matters: we return the first (longest) hit.
+    pub ngram_lens: Vec<usize>,
+    /// Hard cap on spine length. Paper uses 8 — sufficient for typical
+    /// block sizes and avoids running off the end of a match into drift.
+    pub max_extract: usize,
+    /// Minimum extracted length to count as a usable spine. Very short
+    /// spines aren't worth the PLD path (bigram covers 1-token lookahead
+    /// at lower risk); require at least this many continuation tokens.
+    pub min_extract: usize,
+}
+
+impl Default for PldMatcher {
+    fn default() -> Self {
+        Self {
+            ngram_lens: vec![5, 4, 3],
+            max_extract: 8,
+            min_extract: 3,
+        }
+    }
+}
+
+/// Result of a successful PLD lookup.
+#[derive(Debug, Clone)]
+pub struct PldMatch {
+    /// The extracted spine (continuation tokens after the matched suffix).
+    pub tokens: Vec<u32>,
+    /// The suffix length that produced this match (the longest that hit).
+    pub n: usize,
+    /// Number of tried n-gram lengths that agreed on `tokens[0]`. Paper
+    /// §4.3 uses this as part of the bypass-mode confidence signal;
+    /// higher consensus = more reliable spine. Ranges 1..=ngram_lens.len().
+    pub consensus: usize,
+}
+
+impl PldMatcher {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Find a spine continuation for `context`. Returns `None` if no tried
+    /// n-gram length produces a match of length ≥ `self.min_extract`.
+    ///
+    /// For each n in `self.ngram_lens`: take the last-n tokens as the
+    /// suffix, search for its last occurrence earlier in context, and
+    /// extract the `max_extract` tokens that followed it (stopping before
+    /// the suffix itself so we don't include tokens that would be about
+    /// to be re-predicted). Returns the longest-n match with a usable
+    /// spine; consensus counts how many alternate n's produced the same
+    /// first continuation token.
+    pub fn lookup(&self, context: &[u32]) -> Option<PldMatch> {
+        if self.ngram_lens.is_empty() {
+            return None;
+        }
+        // Per-n continuation, collected to compute consensus across lengths.
+        let mut firsts: Vec<u32> = Vec::with_capacity(self.ngram_lens.len());
+        let mut best: Option<(usize, Vec<u32>)> = None; // (n, spine)
+        for &n in &self.ngram_lens {
+            if context.len() <= n {
+                continue;
+            }
+            let suffix_start = context.len() - n;
+            let suffix = &context[suffix_start..];
+            let haystack = &context[..suffix_start];
+            if haystack.len() < n {
+                continue;
+            }
+            // Last occurrence (freshest) of `suffix` in `haystack`.
+            let mut found: Option<usize> = None;
+            for i in (0..=haystack.len() - n).rev() {
+                if &haystack[i..i + n] == suffix {
+                    found = Some(i);
+                    break;
+                }
+            }
+            let start = match found {
+                Some(s) => s,
+                None => continue,
+            };
+            let cont_start = start + n;
+            let cont_end = (cont_start + self.max_extract).min(suffix_start);
+            if cont_end <= cont_start {
+                continue;
+            }
+            let spine: Vec<u32> = context[cont_start..cont_end].to_vec();
+            if spine.len() < self.min_extract {
+                continue;
+            }
+            firsts.push(spine[0]);
+            if best.is_none() {
+                best = Some((n, spine));
+            }
+        }
+
+        let (n, tokens) = best?;
+        let consensus = firsts.iter().filter(|&&t| t == tokens[0]).count();
+        Some(PldMatch {
+            tokens,
+            n,
+            consensus,
+        })
     }
 }
